Various fixes.
Fix: When entering another network, the old identity has to be
subsequently dismissed. But we need to wait a while before then.
See recent changes to documentation of module Qspn.

Fix: Handle rare (but possible) situation where right in the middle
of a duplication some of the qspn-arcs of the old identity are
removed for bad-link.

Also, updates to the tester_?? apps.